Commented Out Python Code for Game Logic
Commented out Python code for game logic because we are using JS instead of Python views to generate the math problems.  Had to modify game.urls.py so the game page would show up and not call the previous Python function.  Updated game.views.py and created a new function that only displays the HTML.